+++ front/pages/w/[wId]/subscription/index.tsx
@@ -1,5 +1,6 @@
 import type { ReactElement } from "react";
 
+import { AdminLayout } from "@app/components/layouts/AdminLayout";
 import { SubscriptionPage } from "@app/components/pages/workspace/subscription/SubscriptionPage";
 import { AppAuthContextLayout } from "@app/components/sparkle/AppAuthContextLayout";
 import type { AppPageWithLayout } from "@app/lib/auth/appServerSideProps";
@@ -15,7 +16,9 @@ PageWithAuthLayout.getLayout = (
   pageProps: AuthContextValue
 ) => {
   return (
-    <AppAuthContextLayout authContext={pageProps}>{page}</AppAuthContextLayout>
+    <AppAuthContextLayout authContext={pageProps}>
+      <AdminLayout>{page}</AdminLayout>
+    </AppAuthContextLayout>
   );
 };
